requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langgraph==0.2.56
langgraph-checkpoint==2.0.8
langgraph-sdk==0.1.43
langsmith==0.1.147
marshmallow==3.23.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.3
orjson==3.10.12
packaging==24.2
propcache==0.2.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
